requirements.txt
dbt-core==1.10.17
dbt-postgres==1.9.1
pandas>=2.0.0
numpy>=1.26.0
psycopg2-binary>=2.9.0
python-dateutil>=2.9.0
PyYAML>=6.0
requests>=2.32.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
